net系のソースにWi-Fi広場関連のソースを加えた

+++ lib/gflib/include/net.h
@@ -265,8 +265,8 @@ typedef struct{
   NetDeleteFriendListCallback friendDeleteFunc;  ///< wifiフレンドリスト削除コールバック
   NetGetDWCFriendDataCallback keyList;   ///< DWC形式の友達リスト	
   NetGetDWCUserDataCallback myUserData;  ///< DWCのユーザデータ（自分のデータ）
-  u16 heapSize;           ///< DWCへのヒープサイズ
-  u16 bDebugServer;        ///< デバック用サーバにつなぐかどうか
+  u32 heapSize;           ///< DWCへのヒープサイズ
+  u32 bDebugServer;        ///< デバック用サーバにつなぐかどうか
 #endif  //GFL_NET_WIFI
   u32 ggid;                 ///< ＤＳでゲームソフトを区別する為のID 任天堂からもらう
   HEAPID baseHeapID;        ///< 元となるHEAPID
